WIP refactor to use model files for db tables, add use strict
