+++ commands/ping.js
@@ -1,10 +1,12 @@
+"use strict";
+
 const { SlashCommandBuilder } = require('@discordjs/builders');
 
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('ping')
 		.setDescription('Replies with Pong!'),
-	async execute(interaction) {
+	async execute(interaction, db) {
 		await interaction.reply('Pong!');
 	},
 };
