+++ commands/popQuiz.js
@@ -1,3 +1,5 @@
+"use strict";
+
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const { Sequelize, DataTypes, Op } = require('sequelize');
 const { MessageActionRow, MessageSelectMenu, Permissions, MessageEmbed } = require('discord.js');
@@ -16,7 +18,7 @@ module.exports = {
                   .setName('topic')
                   .setDescription('Specify the topic for the pop quiz')
                   .setRequired(false)))
-   async execute(interaction){
+   async execute(interaction, db){
       if(interaction.options.getSubcommand() === 'start'){
          const quizObj = JSON.parse(quiz);
          const topic = interaction.options.getString('topic');
