+++ commands/quiz.json
@@ -44,7 +44,7 @@
                   }
                ]
             }
-         },
+         }
       ]
    }
 ]
