+++ events/ready.js
@@ -1,3 +1,5 @@
+"use strict";
+
 module.exports = {
 	name: 'ready',
 	once: true,
