+++ register.js
@@ -1,3 +1,5 @@
+"use strict";
+
 const { REST } = require('@discordjs/rest');
 const { Routes } = require('discord-api-types/v9');
 const fs = require('fs');
